MAINT: renamed module
